requirements.txt
flask
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
numpy==1.17.5
oauthlib==3.1.0
opencv-contrib-python
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow
tensorflow-estimator==2.3.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
tifffile==2020.10.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
